requirements.txt
streamlit
streamlit-option-menu
streamlit-lottie
Pillow